requirements.txt
tensorflow>=2.14.0
numpy
matplotlib
pillow
h5py
scikit-learn
